requirements.txt
numpy~=1.26.3
torch~=2.2.1+cu121
tqdm~=4.66.2
transformers~=4.39.1
pillow~=10.2.0
matplotlib~=3.8.3
torchmetrics~=1.3.2
scikit-learn~=1.4.1.post1
PyYAML~=6.0.1